LayoutTests/platform/mac/accessibility/label-element-with-hidden-control-expected.txt

Test label1 Test label2
This tests that if a label element references a hidden control, it is still visible to accessibility.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS container.childAtIndex('0').role is 'AXRole: AXGroup'
PASS container.childAtIndex('0').childAtIndex(0).role is 'AXRole: AXStaticText'
PASS container.childAtIndex('0').childAtIndex(0).stringValue is 'AXValue: Test label1'
PASS container.childAtIndex('1').role is 'AXRole: AXCheckBox'
PASS container.childAtIndex('1').title is 'AXTitle: Test label2'
PASS successfullyParsed is true

TEST COMPLETE
